backend/requirements.txt
aiohttp
asyncio-nats-client
asyncpg
peewee
